third_party/WebKit/LayoutTests/accessibility/aom-other-actions.html

<!DOCTYPE HTML>
<script src="../resources/gc.js"></script>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="../http/tests/resources/permissions-helper.js"></script>

<!--

Accessibility Object Model
Explainer: https://github.com/WICG/aom/blob/gh-pages/explainer.md
Spec: https://wicg.github.io/aom/spec/

-->
<script>
test(function(t) {
    assert_true(internals.runtimeFlags.accessibilityObjectModelEnabled);
}, "Make sure that Accessibility Object Model is enabled");
</script>

<script>
function enableAccessibilityEventsPermission() {
  return new Promise(function(resolve, reject) {
    PermissionsHelper.setPermission(
        'accessibility-events', 'granted').then(function() {
      // Make sure AXObjectCacheImpl gets the notification too, its
      // listener may fire after this one.
      window.setTimeout(function() {
        resolve();
      }, 0);
    });
  });
}
</script>

<button id="context_menu_target">Context Menu Target</button>

<script>
async_test(function(t) {
  enableAccessibilityEventsPermission().then(function() {
    var target = document.getElementById("context_menu_target");
    var axTarget = accessibilityController.accessibleElementById("context_menu_target");

    var gotDOMEvent = false;
    target.addEventListener("contextmenu", function(event) {
      event.preventDefault();
      gotDOMEvent = true;
    });
    var gotAccessibleEvent = false;
    target.accessibleNode.onaccessiblecontextmenu = function() {
      gotAccessibleEvent = true;
    };
    axTarget.showMenu();

    assert_true(gotAccessibleEvent);
    assert_true(gotDOMEvent);
    t.done();
  });
}, "AccessibleNode.onaccessiblecontextmenu");
</script>

<input type=range min=1 max=5 value=3 id="decrement_target">

<script>
async_test(function(t) {
  enableAccessibilityEventsPermission().then(function() {
    var target = document.getElementById("decrement_target");
    var axTarget = accessibilityController.accessibleElementById("decrement_target");

    var gotAccessibleEvent = false;
    target.accessibleNode.onaccessibledecrement = function() {
      gotAccessibleEvent = true;
    };
    axTarget.decrement();

    assert_true(gotAccessibleEvent);
    assert_equals(target.value, "2");
    t.done();
  });
}, "AccessibleNode.onaccessibledecrement");
</script>

<input type=range min=1 max=5 value=3 id="increment_target">

<script>
async_test(function(t) {
  enableAccessibilityEventsPermission().then(function() {
    var target = document.getElementById("increment_target");
    var axTarget = accessibilityController.accessibleElementById("increment_target");

    var gotAccessibleEvent = false;
    target.accessibleNode.onaccessibleincrement = function() {
      gotAccessibleEvent = true;
    };
    axTarget.increment();

    assert_true(gotAccessibleEvent);
    assert_equals(target.value, "4");
    t.done();
  });
}, "AccessibleNode.onaccessibleincrement");
</script>

<input id="focus_target">

<script>
async_test(function(t) {
  enableAccessibilityEventsPermission().then(function() {
    var target = document.getElementById("focus_target");
    var axTarget = accessibilityController.accessibleElementById("focus_target");

    var gotAccessibleEvent = false;
    target.accessibleNode.onaccessiblefocus = function() {
      gotAccessibleEvent = true;
    };
    axTarget.takeFocus();

    assert_true(gotAccessibleEvent);
    assert_equals(document.activeElement, target);
    t.done();
  });
}, "AccessibleNode.onaccessiblefocus");
</script>

<input id="scroll_target" style="margin-top: 1000px;">

<script>
async_test(function(t) {
  enableAccessibilityEventsPermission().then(function() {
    var target = document.getElementById("scroll_target");
    var axTarget = accessibilityController.accessibleElementById("scroll_target");

    var oldY = window.pageYOffset;

    var gotAccessibleEvent = false;
    target.accessibleNode.onaccessiblescrollintoview = function() {
      gotAccessibleEvent = true;
    };
    axTarget.scrollToMakeVisible();

    assert_true(gotAccessibleEvent);
    assert_true(window.pageYOffset > oldY);
    t.done();
  });
}, "AccessibleNode.onaccessiblescroll");
</script>
